+++ CHANGELOG.md
@@ -1,10 +1,14 @@
+## 1.0.0
+- changed API for checking Tribble readiness
+- Docs to highlight need to call connect() from worker functions
+- Fixed bugs in determining when a Tribble is ready
+- Added example of using large numbers of Tribbles
+
 ## 0.2.0
 - Updated for Dart 3
 
 ## 0.0.1-nullsafety.1
-
 - Null-safe version
 
 ## 0.0.1
-
 - Initial version